add some extra features to CMS like cloning posts and personalizing user login

+++ admin/view_all_posts.php
@@ -18,7 +18,40 @@ if(isset($_POST['checkBoxArray'])){
             $delete_post = mysqli_query($connection, $query);
         break;
 
+        case 'clone':
+            $query = "SELECT * FROM poststb WHERE post_id = $checkBoxPostValueId";
+            $clone_query = mysqli_query($connection, $query);
 
+            if(!$clone_query){
+                die("QUERY FAILED" . mysqli_error($connection));
+
+            }
+            while($row = mysqli_fetch_assoc($clone_query)){
+                $post_category_id = $row['post_category_id'];
+                $post_author = $row['post_author'];
+                $post_title = $row['post_title'];
+                $post_date = $row['post_date'];
+                $post_image = $row['post_image'];
+                $post_content = $row['post_content'];
+                $post_tags = $row['post_tags'];
+                $post_comment_count = $row['post_comment_count'];
+                $post_status = $row['post_status'];
+
+
+                $query = "INSERT INTO poststb (post_category_id, post_author, post_title, ";
+                $query .= "post_date, post_image, post_content, post_tags, post_comment_count, post_status) ";
+                $query .= "VALUES ({$post_category_id}, '{$post_author}', '{$post_title}', '{$post_date}', ";
+                $query .= " '{$post_image}', '{$post_content}', '{$post_tags}', {$post_comment_count}, '{$post_status}')";
+                $clone_insert_query = mysqli_query($connection, $query);
+                if(!$clone_insert_query){
+                    die("Query Failed" . mysqli_error($connection));
+    
+                }
+
+
+            }
+
+        break;
       }
 
     }
@@ -36,6 +69,7 @@ if(isset($_POST['checkBoxArray'])){
         <option value="published">Publish</option>
         <option value="draft">Draft</option>
         <option value="delete">Delete</option>
+        <option value="clone">Clone</option>
     </select>
 </div>
 
@@ -58,6 +92,7 @@ if(isset($_POST['checkBoxArray'])){
             <th>Date</th>
             <th>View Post</th>
             <th>Edit</th>
+            <th>View Count</th>
             <th>Delete</th>
         </tr>
     </thead>
@@ -65,7 +100,7 @@ if(isset($_POST['checkBoxArray'])){
 
 <?php
 
-$query = "SELECT * FROM poststb "; 
+$query = "SELECT * FROM poststb ORDER BY post_id DESC "; 
 $result = mysqli_query($connection, $query);
     while($row = mysqli_fetch_assoc($result)){ 
         $id = $row['post_id'];
@@ -77,6 +112,7 @@ $result = mysqli_query($connection, $query);
         $tags = $row['post_tags'];
         $comment_count = $row['post_comment_count'];
         $date = $row['post_date'];
+        $post_views_count = $row['post_views_count'];
 
         echo "<tr>"; ?>
         <td><input type='checkbox' value='<?php echo $id;  ?>' name='checkBoxArray[]' class='checkBoxes'>
@@ -102,7 +138,8 @@ $result = mysqli_query($connection, $query);
         echo "<td>$date</td>";
         echo "<td><a href='../post.php?post_id={$id}'>View Post</a></td>";
         echo "<td><a href='posts.php?source=edit_posts&edit={$id}'>Edit</a></td>";
-        echo "<td><a href='posts.php?delete={$id}'>Delete</a></td>";
+        echo "<td><a href='posts.php?post_id={$id}&reset={$id}'>$post_views_count</a></td>";
+        echo "<td><a onClick=\"javascript: return confirm('Are You Sure You want to DELETE?');\" href='posts.php?delete={$id}'>Delete</a></td>";
 
    echo "</tr> ";
     }
@@ -120,5 +157,15 @@ $result = mysqli_query($connection, $query);
     }else{
         $delete_id = "";
     }
+    if(isset($_GET['reset'])){
+        $reset_id = $_GET['reset'];
+        $query = "UPDATE poststb SET post_views_count=0 WHERE post_id = {$reset_id}";
+        $reset_query = mysqli_query($connection, $query);
+        header("Location: posts.php");
+    }else{
+        $reset_id = "";
+    }
+
+
 
 ?>
